shift message classes to own file
