+++ python/MANIFEST.in
@@ -1,2 +1,3 @@
 include tox.ini
 recursive-include tests *.py
+recursive-include tests/data *
